refactor: rename mediatailor awsmt
