+++ awsmt/data_source_configuration.go
@@ -1,4 +1,4 @@
-package mediatailor
+package awsmt
 
 import (
 	"context"
