+++ awsmt/data_source_configuration_test.go
@@ -1,4 +1,4 @@
-package mediatailor
+package awsmt
 
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
