+++ main.go
@@ -4,13 +4,13 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 
-	"terraform-provider-mediatailor/mediatailor"
+	"terraform-provider-mediatailor/awsmt"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() *schema.Provider {
-			return mediatailor.Provider()
+			return awsmt.Provider()
 		},
 	})
 }
